requirements.txt
google-cloud-bigquery
google-cloud-secret-manager
google-cloud-firestore
google-cloud-storage
pandas
pyarrow
db-dtypes
pandas-gbq
google-cloud-logging
requests
Flask
Authlib
gunicorn
seleniumbase>=4.30
